+++ src/JitInterface/src/CorInfoImpl.cs
@@ -282,136 +282,6 @@ namespace Internal.JitInterface
             }
         }
 
-        enum RhEHClauseKind
-        {
-            RH_EH_CLAUSE_TYPED = 0,
-            RH_EH_CLAUSE_FAULT = 1,
-            RH_EH_CLAUSE_FILTER = 2
-        }
-
-        private ObjectNode.ObjectData EncodeEHInfo()
-        {
-            var builder = new ObjectDataBuilder();
-            builder.RequireInitialAlignment(1);
-
-            int totalClauses = _ehClauses.Length;
-
-#if READYTORUN
-            for (int i = 0; i < _ehClauses.Length; i++)
-            {
-                ref CORINFO_EH_CLAUSE clause = ref _ehClauses[i];
-                builder.EmitUInt((uint)clause.Flags);
-                builder.EmitUInt(clause.TryOffset);
-                builder.EmitUInt(clause.TryOffset + clause.TryLength);
-                builder.EmitUInt(clause.HandlerOffset);
-                builder.EmitUInt(clause.HandlerOffset + clause.HandlerLength);
-                builder.EmitUInt(clause.ClassTokenOrOffset);
-            }
-            Debug.Assert(builder.CountBytes == 6 * sizeof(uint) * _ehClauses.Length);
-#else
-            // Count the number of special markers that will be needed
-            for (int i = 1; i < _ehClauses.Length; i++)
-            {
-                ref CORINFO_EH_CLAUSE clause = ref _ehClauses[i];
-                ref CORINFO_EH_CLAUSE previousClause = ref _ehClauses[i-1];
-
-                if ((previousClause.TryOffset == clause.TryOffset) && 
-                    (previousClause.TryLength == clause.TryLength) && 
-                    ((clause.Flags & CORINFO_EH_CLAUSE_FLAGS.CORINFO_EH_CLAUSE_SAMETRY) == 0))
-                {
-                    totalClauses++;
-                }
-            }
-
-            builder.EmitCompressedUInt((uint)totalClauses);
-
-            for (int i = 0; i < _ehClauses.Length; i++)
-            {
-                ref CORINFO_EH_CLAUSE clause = ref _ehClauses[i];
-
-                if (i > 0)
-                {
-                    ref CORINFO_EH_CLAUSE previousClause = ref _ehClauses[i-1];
-
-                    // If the previous clause has same try offset and length as the current clause,
-                    // but belongs to a different try block (CORINFO_EH_CLAUSE_SAMETRY is not set),
-                    // emit a special marker to allow runtime distinguish this case.
-                    if ((previousClause.TryOffset == clause.TryOffset) &&
-                        (previousClause.TryLength == clause.TryLength) && 
-                        ((clause.Flags & CORINFO_EH_CLAUSE_FLAGS.CORINFO_EH_CLAUSE_SAMETRY) == 0))
-                    {
-                        builder.EmitCompressedUInt(0);
-                        builder.EmitCompressedUInt((uint)RhEHClauseKind.RH_EH_CLAUSE_FAULT);
-                        builder.EmitCompressedUInt(0);
-                    }
-                }
-
-                RhEHClauseKind clauseKind;
-
-                if (((clause.Flags & CORINFO_EH_CLAUSE_FLAGS.CORINFO_EH_CLAUSE_FAULT) != 0) ||
-                    ((clause.Flags & CORINFO_EH_CLAUSE_FLAGS.CORINFO_EH_CLAUSE_FINALLY) != 0))
-                {
-                    clauseKind = RhEHClauseKind.RH_EH_CLAUSE_FAULT;
-                }
-                else
-                if ((clause.Flags & CORINFO_EH_CLAUSE_FLAGS.CORINFO_EH_CLAUSE_FILTER) != 0)
-                {
-                    clauseKind = RhEHClauseKind.RH_EH_CLAUSE_FILTER;
-                }
-                else
-                {
-                    clauseKind = RhEHClauseKind.RH_EH_CLAUSE_TYPED;
-                }
-
-                builder.EmitCompressedUInt((uint)clause.TryOffset);
-
-                // clause.TryLength returned by the JIT is actually end offset...
-                // https://github.com/dotnet/coreclr/issues/3585
-                int tryLength = (int)clause.TryLength - (int)clause.TryOffset;
-                builder.EmitCompressedUInt((uint)((tryLength << 2) | (int)clauseKind));
-
-                switch (clauseKind)
-                {
-                    case RhEHClauseKind.RH_EH_CLAUSE_TYPED:
-                        {
-                            builder.EmitCompressedUInt(clause.HandlerOffset);
-
-                            var methodIL = (MethodIL)HandleToObject((IntPtr)_methodScope);
-                            var type = (TypeDesc)methodIL.GetObject((int)clause.ClassTokenOrOffset);
-
-                            // Once https://github.com/dotnet/corert/issues/3460 is done, this should be an assert.
-                            // Throwing InvalidProgram is not great, but we want to do *something* if this happens
-                            // because doing nothing means problems at runtime. This is not worth piping a
-                            // a new exception with a fancy message for.
-                            if (type.IsCanonicalSubtype(CanonicalFormKind.Any))
-                                ThrowHelper.ThrowInvalidProgramException();
-
-                            var typeSymbol = _compilation.NodeFactory.NecessaryTypeSymbol(type);
-
-                            RelocType rel = (_compilation.NodeFactory.Target.IsWindows) ?
-                                RelocType.IMAGE_REL_BASED_ABSOLUTE :
-                                RelocType.IMAGE_REL_BASED_REL32;
-
-                            if (_compilation.NodeFactory.Target.Abi == TargetAbi.Jit)
-                                rel = RelocType.IMAGE_REL_BASED_REL32;
-
-                            builder.EmitReloc(typeSymbol, rel);
-                        }
-                        break;
-                    case RhEHClauseKind.RH_EH_CLAUSE_FAULT:
-                        builder.EmitCompressedUInt(clause.HandlerOffset);
-                        break;
-                    case RhEHClauseKind.RH_EH_CLAUSE_FILTER:
-                        builder.EmitCompressedUInt(clause.HandlerOffset);
-                        builder.EmitCompressedUInt(clause.ClassTokenOrOffset);
-                        break;
-                }
-            }
-#endif
-
-            return builder.ToObjectData();
-        }
-
         private void PublishCode()
         {
             var relocs = _relocs.ToArray();
